Use for LMR searches a separate search template parameter.
The idea comes from a measurement of the fail low rate at different searches. For LMR searches we had 19% but for the other searches 62% as fail low rate. So the best code optimizations probably differs (at fail low nodes we have to search all moves if we reach the move loop).

I measured on my computer a speedup of 0.48%:

Engine                          | Nodes/second
./stockfish_lmr                 | 643450.0 +- 13206.98
./stockfish_master              | 640731.5 +- 12344.82

Differences                     | 3101.0 +- 1919.0
Variance of the mean            | 191.9 ( 6.19 %)
Speed up                        | 0.48 %

Bench: 1364759